Fix CI gates for service lifecycle PR
