+++ packages/meta/cli/src/commands/serve.ts
@@ -11,7 +11,7 @@ import { createPatternPermissionBackend } from "@koi/middleware-permissions";
 import { createOpenAICompatAdapter } from "@koi/model-openai-compat";
 import type { ServeFlags } from "../args/serve.js";
 import { resolveApiConfig } from "../env.js";
-import { loadManifestConfig, type ManifestConfig } from "../manifest.js";
+import type { ManifestConfig } from "../manifest.js";
 import type { KoiRuntimeHandle } from "../runtime-factory.js";
 import {
   resolveServiceConfig,
@@ -19,11 +19,16 @@ import {
   serviceHealthUrl,
 } from "../service-lifecycle.js";
 import { ExitCode } from "../types.js";
+import {
+  loadServeManifestConfig,
+  type ManifestConfigLoader,
+  requireServeManifestConfig,
+} from "./serve-manifest.js";
 import { defaultWaitForShutdownSignal } from "./serve-shutdown.js";
 
 export interface ServeDeps {
   readonly waitForShutdownSignal?: (() => Promise<string>) | undefined;
-  readonly loadManifestConfig?: typeof loadManifestConfig | undefined;
+  readonly loadManifestConfig?: ManifestConfigLoader | undefined;
   readonly createRuntime?: (
     service: ServiceConfig,
     flags: ServeFlags,
@@ -182,51 +187,6 @@ async function createServeRuntime(
     ...(manifest.plugins !== undefined ? { plugins: manifest.plugins } : { plugins: [] }),
     ...(manifest.middleware !== undefined ? { manifestMiddleware: manifest.middleware } : {}),
   });
-}
-
-async function requireServeManifestConfig(
-  manifestPath: string,
-  loadConfig?: typeof loadManifestConfig,
-): Promise<ManifestConfig> {
-  const manifest = await loadServeManifestConfig(manifestPath, loadConfig);
-  if (!manifest.ok) throw new Error(manifest.error);
-  return manifest.value;
-}
-
-async function loadServeManifestConfig(
-  manifestPath: string,
-  loadConfig: typeof loadManifestConfig = loadManifestConfig,
-): Promise<ServeResult<ManifestConfig>> {
-  const manifestResult = await loadConfig(manifestPath, {
-    skipAuditValidation: true,
-  });
-  if (!manifestResult.ok) {
-    return { ok: false, error: `invalid manifest - ${manifestResult.error}` };
-  }
-  const unsupported = unsupportedServeManifestFields(manifestResult.value);
-  if (unsupported.length > 0) {
-    return {
-      ok: false,
-      error:
-        "manifest fields are not supported by koi serve yet: " +
-        `${unsupported.join(", ")}. Remove them or run this manifest with a host that honors them.`,
-    };
-  }
-  return { ok: true, value: manifestResult.value };
-}
-
-function unsupportedServeManifestFields(manifest: ManifestConfig): readonly string[] {
-  const fields: string[] = [];
-  if (manifest.backgroundSubprocesses === true) fields.push("backgroundSubprocesses");
-  if (manifest.filesystem !== undefined) fields.push("filesystem");
-  if (manifest.governance !== undefined) fields.push("governance");
-  if (manifest.audit !== undefined) fields.push("audit");
-  if (manifest.network !== undefined) fields.push("network");
-  if (manifest.credentials !== undefined) fields.push("credentials");
-  if (manifest.delegation !== undefined) fields.push("delegation");
-  if (manifest.supervision !== undefined) fields.push("supervision");
-  if (manifest.ace?.enabled === true) fields.push("ace.enabled");
-  return fields;
 }
 
 function createRuntimeGateway(
