+++ packages/meta/cli/src/service-lifecycle.ts
@@ -5,6 +5,13 @@ import { basename, dirname, isAbsolute, join, resolve } from "node:path";
 import { loadConfig } from "@koi/config";
 import { loadManifestConfig } from "./manifest.js";
 import { resolveManifestPath } from "./resolve-manifest-path.js";
+import {
+  escapeXml,
+  parseKeyValueLines,
+  parsePsOutput,
+  shellQuote,
+  systemdQuote,
+} from "./service-lifecycle-format.js";
 
 export type ServicePlatform = "linux" | "darwin";
 export type ServiceRestart = "on-failure" | "always" | "no";
@@ -739,49 +746,6 @@ async function fileReadable(path: string): Promise<boolean> {
   }
 }
 
-function parseKeyValueLines(output: string): ReadonlyMap<string, string> {
-  const props = new Map<string, string>();
-  for (const line of output.split("\n")) {
-    const idx = line.indexOf("=");
-    if (idx === -1) continue;
-    props.set(line.slice(0, idx), line.slice(idx + 1).trim());
-  }
-  return props;
-}
-
-function parsePsOutput(output: string): {
-  readonly uptimeMs: number | undefined;
-  readonly memoryBytes: number | undefined;
-} {
-  const parts = output.trim().split(/\s+/);
-  const rssKb = Number.parseInt(parts[0] ?? "", 10);
-  return {
-    memoryBytes: Number.isNaN(rssKb) ? undefined : rssKb * 1024,
-    uptimeMs: parseElapsed(parts[1] ?? ""),
-  };
-}
-
-function parseElapsed(value: string): number | undefined {
-  let days = 0;
-  let rest = value;
-  const dayMatch = /^(\d+)-(.+)$/.exec(value);
-  if (dayMatch !== null) {
-    days = Number.parseInt(dayMatch[1] ?? "0", 10);
-    rest = dayMatch[2] ?? "";
-  }
-  const parts = rest.split(":").map((p) => Number.parseInt(p, 10));
-  if (parts.some(Number.isNaN)) return undefined;
-  if (parts.length === 2) {
-    const [minutes, seconds] = parts as [number, number];
-    return ((days * 24 * 60 + minutes) * 60 + seconds) * 1000;
-  }
-  if (parts.length === 3) {
-    const [hours, minutes, seconds] = parts as [number, number, number];
-    return (((days * 24 + hours) * 60 + minutes) * 60 + seconds) * 1000;
-  }
-  return undefined;
-}
-
 async function bestEffort(exec: ExecFn, argv: readonly string[]): Promise<void> {
   await exec(argv);
 }
@@ -805,22 +769,4 @@ async function execCommand(argv: readonly string[]): Promise<ExecResult> {
     proc.exited,
   ]);
   return { exitCode, stdout, stderr };
-}
-
-function systemdQuote(value: string): string {
-  if (/^[A-Za-z0-9_@%+=:,./-]+$/.test(value)) return value.replace(/%/g, "%%");
-  return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"').replace(/%/g, "%%")}"`;
-}
-
-function shellQuote(value: string): string {
-  return `'${value.replace(/'/g, "'\\''")}'`;
-}
-
-function escapeXml(value: string): string {
-  return value
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&apos;");
 }
